requirements.txt
fastapi==0.125.0
uvicorn
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.11.0
supabase==2.27.0
httpx==0.28.1
loguru==0.7.3
python-dotenv==1.2.1
anyio==4.12.0

numpy==2.0.2
pillow==11.3.0
scipy==1.13.1
scikit-image==0.24.0
tifffile==2024.8.30

cellpose==2.0.2
torch==2.8.0

psutil
matplotlib==3.9.4
